Incorporate the flushless producer ability inside the main Kafka connector task
